requirements.txt
asgiref==3.9.1
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==46.0.1
dj-database-url==3.0.1
Django==5.2.6
django-allauth==65.11.2
django-anymail==13.1
django-cors-headers==4.9.0
django-extensions==4.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
MarkupSafe==3.0.2
packaging==25.0
pillow==12.0.0
psycopg2-binary==2.9.11
pycparser==2.23
PyJWT==2.10.1
python-dotenv==1.1.1
python-http-client==3.3.7
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
resend==2.15.0
sendgrid==3.6.5
sendgrid-django==4.2.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
Werkzeug==3.1.3
python-decouple

